ステップ13の検証用コードを正式コードに変更
